+++ wradlib/tests/test_classify.py
@@ -15,7 +15,18 @@ from . import get_wradlib_data_file, requires_gdal, requires_h5py, requires_netc
 
 
 def test_filter_gabella_a():
-    pass
+    filename = get_wradlib_data_file("misc/polar_dBZ_fbg.gz")
+    data = np.loadtxt(filename)
+    clutter = classify.filter_gabella_a(data, wsize=5, tr1=6.0)
+
+    site = (47.875, 8.004, 1489.6)
+    r = np.arange(0, 128000, 1000) + 500
+    az = np.arange(0, 360, 1) + 0.5
+    fbg = georef.create_xarray_dataarray(data, r=r, phi=az, site=site)
+
+    clutterx = fbg.wrl.classify.filter_gabella_a(wsize=5, tr1=6.0)
+
+    np.testing.assert_array_equal(clutterx.values, clutter)
 
 
 def test_filter_window_distance():
@@ -25,6 +36,15 @@ def test_filter_window_distance():
     img[20, :] = 5  # spike
     img[9:12, 4:7] = 11  # precip field
     img[15:17, 5:7] = np.nan  # nans
+    img_arr = xr.DataArray(
+        img,
+        dims=["azimuth", "range"],
+        coords=[
+            np.linspace(0, 360, 36, endpoint=False),
+            np.linspace(0, 2500, 10, endpoint=False),
+        ],
+    )
+
     cl = img.copy()
     cl[img > 0] = True
     cl[img == 11] = False
@@ -36,17 +56,42 @@ def test_filter_window_distance():
     np.set_printoptions(precision=3)
     assert (result == cl).all()
 
+    similarx = img_arr.wrl.classify.filter_window_distance(fsize=300, tr1=4)
+    resultx = similarx < 0.3
+    assert (resultx.values == cl).all()
+
 
 def test_filter_gabella():
     filename = get_wradlib_data_file("misc/polar_dBZ_fbg.gz")
     data = np.loadtxt(filename)
-    classify.filter_gabella(data, wsize=5, thrsnorain=0.0, tr1=6.0, n_p=8, tr2=1.3)
+    clutter = classify.filter_gabella(
+        data, wsize=5, thrsnorain=0.0, tr1=6.0, n_p=8, tr2=1.3
+    )
+
+    site = (47.875, 8.004, 1489.6)
+    r = np.arange(0, 128000, 1000) + 500
+    az = np.arange(0, 360, 1) + 0.5
+    fbg = georef.create_xarray_dataarray(data, r=r, phi=az, site=site)
+
+    clutterx = fbg.wrl.classify.filter_gabella(
+        wsize=5, thrsnorain=0.0, tr1=6.0, n_p=8, tr2=1.3
+    )
+
+    np.testing.assert_array_equal(clutterx.values, clutter)
 
 
 def test_histo_cut():
     filename = get_wradlib_data_file("misc/annual_rainfall_fbg.gz")
     yearsum = np.loadtxt(filename)
-    classify.histo_cut(yearsum)
+    clutter = classify.histo_cut(yearsum)
+
+    site = (47.875, 8.004, 1489.6)
+    r = np.arange(0, 128000, 1000) + 500
+    az = np.arange(0, 360, 1) + 0.5
+    fbg = georef.create_xarray_dataarray(yearsum, r=r, phi=az, site=site)
+    clutterx = fbg.wrl.classify.histo_cut()
+
+    np.testing.assert_array_equal(clutterx.values, clutter)
 
 
 @pytest.fixture()
@@ -63,12 +108,26 @@ def fuzzy_data():
     dat["dop"], attrs_dop = io.read_edge_netcdf(dopfile)
     dat["zdr"], attrs_zdr = io.read_edge_netcdf(zdrfile)
     dat["map"] = io.from_hdf5(mapfile)[0][0]
-    yield dat
+
+    datx = {k: (["azimuth", "range"], v) for k, v in dat.items()}
+    az, rng = dat["rho"].shape
+    swp = xr.Dataset(datx, coords={"azimuth": np.arange(az), "range": np.arange(rng)})
+    swp = swp.assign_coords(
+        dict(
+            longitude=7,
+            latitude=53,
+            altitude=0,
+            elevation=1,
+            sweep_mode="azimuth_surveillance",
+        )
+    )
+    yield dat, swp
 
 
 @requires_netcdf
 @requires_h5py
 def test_classify_echo_fuzzy(fuzzy_data):
+    dat, swp = fuzzy_data
     weights = {
         "zdr": 0.4,
         "rho": 0.4,
@@ -77,7 +136,7 @@ def test_classify_echo_fuzzy(fuzzy_data):
         "dop": 0.1,
         "map": 0.5,
     }
-    prob, mask = classify.classify_echo_fuzzy(fuzzy_data, weights=weights)
+    prob, mask = classify.classify_echo_fuzzy(dat, weights=weights)
     np.testing.assert_array_equal(
         prob[0, :4],
         np.array(
@@ -89,6 +148,12 @@ def test_classify_echo_fuzzy(fuzzy_data):
             ]
         ),
     )
+
+    moments = dict(rho="rho", phi="phi", dop="dop", zdr="zdr", map="map")
+    probx, maskx = swp.wrl.classify.classify_echo_fuzzy(moments, weights=weights)
+
+    np.testing.assert_array_almost_equal(probx.values, prob)
+    np.testing.assert_array_equal(maskx.values, mask)
 
 
 @pytest.fixture()